Rename for case sensitive file system - 2
